timing fix

+++ include/uart.h
@@ -14,7 +14,7 @@
 int openSerialPort(const std::string& portname);
 bool configureSerialPort(int fd, int speed);
 int sendCommand(int fd,const std::string& cmd);
-std::vector<std::string> readResponse(int fd, int timeout_sec);
+std::vector<std::string> readResponse(int fd, double timeout_sec);
 int uart_init(const std::string& portname, int br);
 int get_baudrate(int);
 #endif
